refactor(allegro): apply tech-review fixes for #584
Move `allegro-scheduler-tasks.ts` (and its spec) under
`infrastructure/scheduler/` so the file sits inside one of the four
sanctioned hexagonal layers (`engineering-standards.md` §"Standard
Module Structure") — the previous top-level `src/scheduler/` folder
broke that convention.

Drive-by:
- build the offers-sync payload conditionally so `cursorKey` is only
  included when `feedType === 'events'` (no more literal `undefined`
  values on the wire)
- normalize the registration log's scope string `'none'` → `'unknown'`
  so it matches the same code path's execute-time log
